EC/backend/API/text-to-image/requirements.txt
accelerate==0.26.1
altair==5.2.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
diffusers==0.26.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.47.2
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
invisible-watermark==0.2.0
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.2
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.31.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
